+++ guillotina/utils/__init__.py
@@ -17,6 +17,7 @@ from .crypto import secure_passphrase  # noqa
 from .misc import apply_coroutine  # noqa
 from .misc import get_current_request  # noqa
 from .misc import get_random_string  # noqa
+from .misc import get_schema_validator  # noqa
 from .misc import get_url  # noqa
 from .misc import lazy_apply  # noqa
 from .misc import list_or_dict_items  # noqa